- moved all Windows-specific portability header files to the /win/ directory --> sys_time.h has moved. - tweak CMake scripts to write the current config in config.h   + moved and augmented config.h to config.h.in - fixed a couple of integer size related MSVC compiler warnings - make sure all sources pick up the currently active config.h file, whether the one produced by the local CMake or the one from the monolithic build environment.

+++ unicode_gen.c
@@ -92,7 +92,7 @@ const char *get_field_buf(char *buf, size_t buf_size, const char *p, int n)
     p = get_field(p, n);
     q = buf;
     while (*p != ';' && *p != '\0') {
-        if ((q - buf) < buf_size - 1)
+        if ((size_t)(q - buf) < buf_size - 1)
             *q++ = *p;
         p++;
     }
